+++ src/modules/module-loopback.c
@@ -248,7 +248,7 @@ static void capture_destroy(void *d)
 
 static void recalculate_delay(struct impl *impl)
 {
-	uint32_t target = impl->rate * impl->target_delay, cdelay, pdelay;
+	uint32_t target = (uint32_t)(impl->rate * impl->target_delay), cdelay, pdelay;
 	uint32_t delay, w;
 	struct pw_time pwt;
 
@@ -435,7 +435,7 @@ static void param_format_changed(struct impl *impl, const struct spa_pod *param,
 static void recalculate_buffer(struct impl *impl)
 {
 	if (impl->target_delay > 0.0f && impl->channels > 0 && impl->rate > 0) {
-		uint32_t delay = impl->rate * impl->target_delay;
+		uint32_t delay = (uint32_t)(impl->rate * impl->target_delay);
 		void *data;
 
 		impl->buffer_size = (delay + (1u<<15)) * 4;
